Update main.css
Changed the *-perspective settings to 479px to fix the distortion appearing in browsers. I've tested this in Firefox, IE11 and Chrome on Win7, Safari on Mac OS X El Capitan, and Edge on Windows 10. 
It's live on the websites of the company I work for http://www.wardandpartners.co.uk/black-friday-offer/ I'm not trying to sell you anything - honest!
Fix suggested by https://github.com/vilius-g