Remove hasEventsAttached variable
and replace it with a simple boolean that ensures that at
least two ticks are handled. This is necessary, so that
event listeners registered after the invocation can still
handle an event.

The change in general allows events to be handle by listeners
attached to parent elements of the countdown element.